backend_python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
solana==0.30.2
solders==0.18.1
requests-oauthlib==1.3.1
PyJWT==2.8.0
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
